+++ rana_qgis_plugin/rana_qgis_plugin.py
@@ -404,6 +404,13 @@ class RanaQgisPlugin:
             self.rana_browser.create_model_selected_with_revision.connect(
                 self.loader.create_schematisation_revision_3di_model
             )
+            self.rana_browser.export_gpkg_selected.connect(self.rana_browser.disable)
+            self.rana_browser.export_gpkg_selected.connect(
+                self.loader.export_schematisation_from_file
+            )
+            self.rana_browser.export_gpkg_revision_selected.connect(
+                self.loader.export_schematisation_revision
+            )
             self.rana_browser.delete_model_selected.connect(
                 self.loader.delete_schematisation_revision_3di_model
             )
@@ -439,6 +446,10 @@ class RanaQgisPlugin:
             self.loader.raster_style_finished.connect(self.rana_browser.enable)
             self.loader.raster_style_finished.connect(self.rana_browser.refresh)
             self.loader.raster_style_failed.connect(self.rana_browser.enable)
+            self.loader.export_gpkg_finished.connect(self.rana_browser.enable)
+            self.loader.export_gpkg_finished.connect(
+                self.rana_browser.return_to_file_browser
+            )
             self.loader.publication_style_finished.connect(
                 self.rana_browser.publication_view.update_after_save_styles
             )
@@ -452,9 +463,7 @@ class RanaQgisPlugin:
             self.loader.schematisation_upload_failed.connect(self.rana_browser.enable)
             self.loader.folder_created.connect(self.rana_browser.refresh)
             self.loader.model_deleted.connect(self.rana_browser.refresh)
-            self.loader.file_deleted.connect(
-                self.rana_browser.refresh_after_file_delete
-            )
+            self.loader.file_deleted.connect(self.rana_browser.return_to_file_browser)
             self.loader.rename_aborted.connect(self.rana_browser.refresh)
             self.loader.rename_finished.connect(
                 self.rana_browser.refresh_after_file_rename